Add duke package
