+++ src/main/java/duke/Event.java
@@ -1,15 +1,17 @@
+package duke;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private LocalDate dateTime;
-    public Deadline(String description, String dateTime) throws DukeException {
+    public Event(String description, String dateTime) throws DukeException {
         super(description);
         this.dateTime = LocalDate.parse(dateTime);
     }
     public String toString() {
         String dateString = dateTime.format(DateTimeFormatter.ofPattern("MMM dd yyyy"));
-        return String.format("[D][%s] %s (by: %s)",
+        return String.format("[E][%s] %s (at: %s)",
                 getStatusIcon(), description, dateString);
     }
 
@@ -19,6 +21,6 @@ public class Deadline extends Task {
 
     @Override
     public TaskType getType() {
-        return TaskType.DEADLINE;
+        return TaskType.EVENT;
     }
 }
