+++ src/main/java/duke/Storage.java
@@ -1,12 +1,9 @@
-import java.nio.file.Path;
-import java.nio.file.Files;
-import java.nio.file.Paths;
+package duke;
+
 import java.util.ArrayList;
 import java.util.Scanner;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
-import java.io.IOException;
 
 public class Storage {
     private String filePath;
