+++ src/main/java/duke/TriConsumer.java
@@ -1,3 +1,5 @@
+package duke;
+
 @FunctionalInterface
 public interface TriConsumer {
     void execute(TaskList taskList, Ui ui, Storage storage) throws DukeException, Exception;
